+++ .phan/baseline.php
@@ -8,25 +8,9 @@
  * (can be combined with --load-baseline)
  */
 return [
-	// # Issue statistics:
-	// PhanTypeMismatchArgumentInternal : 10+ occurrences
-	// SecurityCheck-XSS : 6 occurrences
-	// MediaWikiNoBaseException : 2 occurrences
-	// PhanTypeMismatchArgument : 2 occurrences
-	// PhanUndeclaredClassMethod : 2 occurrences
-	// PhanUndeclaredTypeParameter : 2 occurrences
-	// PhanParamSignatureMismatch : 1 occurrence
-	// PhanParamTooMany : 1 occurrence
-	// PhanTypeInvalidLeftOperandOfAdd : 1 occurrence
-	// PhanTypeMismatchReturn : 1 occurrence
-	// PhanTypeMismatchReturnProbablyReal : 1 occurrence
-
+	// This baseline has no suppressions
 	// Currently, file_suppressions and directory_suppressions are the only supported suppressions
 	'file_suppressions' => [
-		'src/EditWarning.php' => ['PhanParamTooMany', 'PhanTypeInvalidLeftOperandOfAdd', 'PhanTypeMismatchArgumentInternal', 'PhanTypeMismatchReturn', 'PhanTypeMismatchReturnProbablyReal', 'PhanUndeclaredClassMethod', 'PhanUndeclaredTypeParameter'],
-		'src/EditWarningApi.php' => ['PhanTypeMismatchArgument'],
-		'src/EditWarningMessage.php' => ['MediaWikiNoBaseException', 'SecurityCheck-XSS'],
-		'src/EditWarningMsg.php' => ['PhanParamSignatureMismatch', 'SecurityCheck-XSS'],
 	],
 	// 'directory_suppressions' => ['src/directory_name' => ['PhanIssueName1', 'PhanIssueName2']] can be manually added if needed.
 	// (directory_suppressions will currently be ignored by subsequent calls to --save-baseline, but may be preserved in future Phan releases)